fix(home): tidy up homepage and fix menu links
