+++ frontend/src/components/ui/shadcn-io/navbar-01/index.tsx
@@ -212,7 +212,14 @@ export const Navbar01 = React.forwardRef<HTMLElement, Navbar01Props>(
                     {navigationLinks.map((link, index) => (
                       <NavigationMenuItem key={index}>
                         <button
-                          onClick={(e) => e.preventDefault()}
+                          onClick={(e) => {
+                            e.preventDefault();
+                            link.href.startsWith("#")
+                              ? document
+                                  ?.querySelector(link.href)
+                                  ?.scrollIntoView({ behavior: "smooth" })
+                              : navigate(link.href);
+                          }}
                           className={cn(
                             "group inline-flex h-9 w-max items-center justify-center rounded-md px-4 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground focus:outline-none disabled:pointer-events-none disabled:opacity-50 cursor-pointer no-underline",
                             link.active
