+++ frontend/src/pages/homepage/components/about-section.tsx
@@ -1,52 +1,27 @@
 "use client";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Check, File, Microscope, UserCircle } from "lucide-react";
 
-import { Card, CardContent } from "@/components/ui/card";
-import { CardDecorator } from "@/components/ui/card-decorator";
-import {
-  Search,
-  Trophy,
-  ChartBar,
-  Workflow,
-  Pen,
-  BookOpen,
-} from "lucide-react";
-
-const values = [
+const items = [
   {
-    icon: BookOpen,
-    title: "View Knowledge Bricks",
-    description:
-      "Beautiful, interactive display of scientific nanopublications with automatic template fetching.",
+    label: "Structured:",
+    text: "Clear assertion, provenance, and publication info",
   },
   {
-    icon: Pen,
-    title: "Create Knowledge Bricks",
-    description:
-      "Transform research findings into FAIR nanopublications using intuitive templates.",
+    label: "Linked:",
+    text: "Connected to other knowledge using standard vocabularies",
   },
   {
-    icon: Workflow,
-    title: "Connect Research",
-    description:
-      "Link findings across studies, creating a connected knowledge graph.",
+    label: "Attributed:",
+    text: "Properly credited to its creator",
   },
   {
-    icon: Trophy,
-    title: "Get Credit",
-    description:
-      "Proper attribution and credit for every contribution through ORCID integration.",
+    label: "Verifiable:",
+    text: "Cryptographically signed and trustworthy",
   },
   {
-    icon: Search,
-    title: "Discover Knowledge",
-    description:
-      "PSearch and find relevant nanopublications across the scientific community.",
-  },
-  {
-    icon: ChartBar,
-    title: "Analyze Impact",
-    description:
-      "Track how your knowledge bricks are used and built upon by others.",
+    label: "Permanent:",
+    text: "Immutable and permanently accessible",
   },
 ];
 
@@ -54,35 +29,75 @@ export function AboutSection() {
   return (
     <section id="about" className="py-24 sm:py-32">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8">
-        {/* Section Header */}
-        <div className="mx-auto max-w-4xl text-center mb-16">
-          <h2 className="text-3xl font-bold tracking-tight sm:text-4xl mb-6">
-            Key Features
-          </h2>
-          <p className="text-lg text-muted-foreground mb-8">
-            Built for Open Science.
-          </p>
-        </div>
+        <div className="grid gap-8 lg:grid-cols-2">
+          {/* Info */}
+          <div className="lg:col-span-1 order-1 lg:order-1">
+            <h2 className="text-3xl font-bold tracking-tight sm:text-4xl mb-4">
+              What are Nanopublications?
+            </h2>
+            <h3 className="mb-4">
+              Nanopublications are the smallest units of publishable
+              information. Each nanopublication is:
+            </h3>
+            <ul role="list" className="space-y-3 text-sm">
+              {items.map((feature) => (
+                <li className="flex items-center gap-3">
+                  <Check
+                    className="text-muted-foreground size-4 flex-shrink-0"
+                    strokeWidth={2.5}
+                  />
+                  <span>
+                    <strong>{feature.label}</strong> {feature.text}
+                  </span>
+                </li>
+              ))}
+            </ul>
+          </div>
 
-        {/* Modern Values Grid with Enhanced Design */}
-        <div className="grid grid-cols-1 gap-x-8 gap-y-12 sm:grid-cols-2 xl:grid-cols-3 mb-12">
-          {values.map((value, index) => (
-            <Card key={index} className="group shadow-xs py-2">
-              <CardContent className="p-8">
-                <div className="flex flex-col items-center text-center">
-                  <CardDecorator>
-                    <value.icon className="h-6 w-6" aria-hidden />
-                  </CardDecorator>
-                  <h3 className="mt-6 font-medium text-balance">
-                    {value.title}
-                  </h3>
-                  <p className="text-muted-foreground mt-3 text-sm">
-                    {value.description}
-                  </p>
-                </div>
+          {/* Side boxes */}
+          <div className="space-y-6 order-2 lg:order-2">
+            <Card className="hover:shadow-md transition-shadow cursor-pointer">
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2">
+                  <File className="h-5 w-5 text-primary" />
+                  Assertion
+                </CardTitle>
+              </CardHeader>
+              <CardContent>
+                <p className="text-muted-foreground mb-3">
+                  The actual claim or statement.
+                </p>
               </CardContent>
             </Card>
-          ))}
+
+            <Card className="hover:shadow-md transition-shadow cursor-pointer">
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2">
+                  <Microscope className="h-5 w-5 text-primary" />
+                  Provenance
+                </CardTitle>
+              </CardHeader>
+              <CardContent>
+                <p className="text-muted-foreground mb-3">
+                  Where did this come from?
+                </p>
+              </CardContent>
+            </Card>
+
+            <Card className="hover:shadow-md transition-shadow cursor-pointer">
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2">
+                  <UserCircle className="h-5 w-5 text-primary" />
+                  Publication Info
+                </CardTitle>
+              </CardHeader>
+              <CardContent>
+                <p className="text-muted-foreground mb-3">
+                  Who, when, and how?
+                </p>
+              </CardContent>
+            </Card>
+          </div>
         </div>
       </div>
     </section>
