+++ frontend/src/pages/homepage/components/cta-section.tsx
@@ -5,7 +5,7 @@ import { ArrowRight } from "lucide-react";
 
 export function CtaSection() {
   return (
-    <section id="about" className="py-24 sm:py-32">
+    <section id="cta-bottom" className="py-24 sm:py-32">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8">
         {/* Section Header */}
         <div className="mx-auto max-w-4xl text-center mb-16">
