+++ frontend/src/pages/homepage/page.tsx
@@ -1,9 +1,9 @@
 import { Navbar01 } from "@/components/ui/shadcn-io/navbar-01";
 import { HeroSection } from "./components/hero-section";
-import { AboutSection } from "./components/about-section";
+import { FeaturesSection } from "./components/features-section";
 import { CtaSection } from "./components/cta-section";
 import { LandingFooter } from "./components/footer";
-import { QuestionSection } from "./components/question-section";
+import { AboutSection } from "./components/about-section";
 import { TeamSection } from "./components/team-section";
 
 export function HomePage() {
@@ -15,9 +15,9 @@ export function HomePage() {
       </div>
       <HeroSection />
 
-      <AboutSection />
+      <FeaturesSection />
 
-      <QuestionSection />
+      <AboutSection />
 
       <TeamSection />
